+++ tests/Validators/TokenValidatorTest.php
@@ -84,7 +84,7 @@ class TokenValidatorTest extends AbstractTestCase
         $this->validator->check($token);
     }
 
-    public function dataProviderMalformedTokens()
+    public static function dataProviderMalformedTokens()
     {
         return [
             ['one.two.'],
@@ -97,7 +97,7 @@ class TokenValidatorTest extends AbstractTestCase
         ];
     }
 
-    public function dataProviderTokensWithWrongSegmentsNumber()
+    public static function dataProviderTokensWithWrongSegmentsNumber()
     {
         return [
             ['one.two'],