+++ CHANGELOG.md
@@ -8,6 +8,9 @@ You can find and compare releases at the GitHub release page.
 
 ## [Unreleased]
 
+### Added
+- Different TTL configurations for each guard
+
 ## [2.0.0] 2022-09-08
 - No changes to 2.0.0-RC1