+++ CHANGELOG.md
@@ -11,6 +11,12 @@ You can find and compare releases at the GitHub release page.
 ### Added
 - Different TTL configurations for each guard
 - lcobucci/jwt: add array support for `aud` claim
+- Laravel 11 support
+
+## [2.1.0] 2023-02-17
+
+### Added
+- Laravel 10 support
 
 ## [2.0.0] 2022-09-08
 - No changes to 2.0.0-RC1